Move user role module into models/concerns
