+++ app/models/user.rb
@@ -12,7 +12,7 @@ class User < ActiveRecord::Base
   end
 
   include Overridable
-  include Role
+  include ::Users::Roles
   include NUCore::Database::WhereIdsIn
 
   # ldap_authenticatable is included via a to_prepare hook if ldap is enabled
